Fix button padding and add new button_toggle_icon widget
